% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{AcidShiny-package}
\alias{AcidShiny}
\alias{AcidShiny-package}
\title{AcidShiny}
\description{
Shiny app development functions for server and UI consistency.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/acidshiny/}
  \item \url{https://github.com/acidgenomics/r-acidshiny/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-acidshiny/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
