% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderAcidDataTable.R
\name{renderAcidDataTable}
\alias{renderAcidDataTable}
\title{Render table output with the JavaScript library DataTables using DT}
\usage{
renderAcidDataTable(expr, env = parent.frame())
}
\arguments{
\item{expr}{\code{data.frame}.}

\item{env}{\code{environment}.}
}
\description{
Render table output with the JavaScript library DataTables using DT
}
\note{
Updated 2021-04-07.
}
\examples{
data(mtcars, package = "datasets")

if (interactive()) {
    expr <- mtcars
    x <- renderAcidDataTable(expr)
    class(x)
}
}
